move AutoClassify to util
